Audio Data
This is another #AudioData classification project aiming to label the song genre as Hip-Hop or Rock. Two separate datasets of features and labels were merged for #classification purposes in later stages. The next step involves the computation of #correlations  between the features where only the features with weak correlations will be picked for further processing, as it tries to avoid the excess correlations that only bring too much noise to regression. After removing one problematic feature and label and data #scaling and #standardizing, the #PCA will extract 8 remaining explained variations, but no clear #elbow on a #ScreePlot is visible. The solution is in plotting the #CumulativeVariation which shows that the first 6 components explain 85% of the variation. Now we have the data prepared for classification with #DecisionTree and #LogisticRegression. Both classifications have a high accuracy of about 0.87 and good #precision, #recall, and #f1score metrics in predicting the Rock genre but showing some variation regarding HipHop. We'll try to remedy this issue by balancing the data in such a way that there is the same number of both rock and hip-hop songs.